BUG: Redo DockerHub image tagging yet again.
Can't easily use the mutliple-tag variant of DockerHub triggers because the tag is actually the full tag with the commas. Parsing it is possible, but then need to know which is the full tag to split to get labels in the build script. Easier to just use a post-push script and only trigger and build the full x-y tag.